+++ src/splits_to_tree.cpp
@@ -78,7 +78,7 @@ IntegerMatrix splits_to_edge(const RawMatrix splits, const IntegerVector nTip) {
     for (int16 bin = x.n_bins; bin--; ) {
       const splitbit chunk = x.state[split_order[split]][bin];
       for (int16 bin_tip = SL_BIN_SIZE; bin_tip--; ) {
-        const int16 tip = bin_tip + (bin * SL_BIN_SIZE);
+        const int16 tip = bin_tip + int16(bin * SL_BIN_SIZE);
         if (chunk & powers_of_two[bin_tip]) {
           insert_ancestor(tip, &next_node, parent, patriarch);
         }
